mcp-service/requirements.txt
# MCP服务核心依赖
fastapi==0.104.1
uvicorn==0.24.0.post1
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.1.1

# 日志和工具
python-multipart==0.0.6
